refactor(portfolio-deploy): vstorage-outdated
